pyproject.toml
[project]
name = "gdg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "chromedriver-autoinstaller>=0.6.4",
    "fastapi>=0.115.12",
    "geopy>=2.4.1",
    "google-generativeai>=0.8.5",
    "googlemaps>=4.10.0",
    "httpx>=0.28.1",
    "joblib>=1.5.0",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.0.10",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "numpy>=2.1.3",
    "openrouteservice>=2.3.3",
    "pandas>=2.2.3",
    "prophet>=1.1.6",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "selenium>=4.33.0",
    "uvicorn>=0.34.2",
    "webdriver-manager>=4.0.2",
]

[tool.taskipy.tasks]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
test = "uv run app/main.py"

[dependency-groups]
dev = [
    "taskipy>=1.14.1",
]

